// src/models/DescCen4.ts
export interface DescCen4 {
  IdDescCen4: number;
  nameDescCen4: string | null;
  codeDescCen4: string | null;
}

export interface DescCen4Create {
  nameDescCen4: string | null;
  codeDescCen4: string | null;
}

export interface DescCen4Update {
  nameDescCen4: string | null;
  codeDescCen4: string | null;
}